[SDSAET-38] Move JsErrorLog to models-api module
