+++ api/jobs-api/pom.xml
@@ -43,6 +43,10 @@
       <groupId>com.cognifide.aet</groupId>
       <artifactId>datastorage-api</artifactId>
     </dependency>
+    <dependency>
+      <groupId>com.cognifide.aet</groupId>
+      <artifactId>models-api</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>org.apache.commons</groupId>
