+++ api/models-api/src/main/java/com/cognifide/aet/models/ErrorType.java
