+++ core/jobs/src/main/java/com/cognifide/aet/job/common/collectors/jserrors/JsErrorsCollector.java
@@ -18,9 +18,9 @@ package com.cognifide.aet.job.common.collectors.jserrors;
 import com.cognifide.aet.communication.api.metadata.CollectorStepResult;
 import com.cognifide.aet.job.api.collector.CollectorJob;
 import com.cognifide.aet.job.api.collector.CollectorProperties;
-import com.cognifide.aet.job.api.collector.JsErrorLog;
 import com.cognifide.aet.job.api.collector.WebCommunicationWrapper;
 import com.cognifide.aet.job.api.exceptions.ProcessingException;
+import com.cognifide.aet.models.jserrors.JsErrorLog;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import java.util.Map;
 import java.util.Set;
