+++ core/jobs/src/main/java/com/cognifide/aet/job/common/groupers/jserrors/JsErrorsGrouper.java
@@ -19,10 +19,10 @@ package com.cognifide.aet.job.common.groupers.jserrors;
 import com.cognifide.aet.communication.api.JobStatus;
 import com.cognifide.aet.communication.api.job.GrouperJobData;
 import com.cognifide.aet.communication.api.job.GrouperResultData;
-import com.cognifide.aet.job.api.collector.JsErrorLog;
 import com.cognifide.aet.job.api.grouper.GrouperJob;
 import com.cognifide.aet.job.common.groupers.GroupingAlgorithm;
 import com.cognifide.aet.job.common.groupers.GroupingException;
+import com.cognifide.aet.models.jserrors.JsErrorLog;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.cognifide.aet.vs.DBKey;
 import com.google.common.base.Strings;
