+++ core/jobs/src/test/java/com/cognifide/aet/job/common/groupers/jserrors/JsErrorsGrouperTest.java
@@ -29,9 +29,9 @@ import com.cognifide.aet.communication.api.job.GrouperJobData;
 import com.cognifide.aet.communication.api.job.GrouperResultData;
 import com.cognifide.aet.communication.api.metadata.ComparatorStepResult;
 import com.cognifide.aet.communication.api.metadata.ComparatorStepResult.Status;
-import com.cognifide.aet.job.api.collector.JsErrorLog;
 import com.cognifide.aet.job.common.groupers.GroupingAlgorithm;
 import com.cognifide.aet.job.common.groupers.GroupingException;
+import com.cognifide.aet.models.jserrors.JsErrorLog;
 import com.cognifide.aet.vs.ArtifactsDAO;
 import com.cognifide.aet.vs.DBKey;
 import com.cognifide.aet.vs.SimpleDBKey;
