+++ core/worker/pom.xml
@@ -48,6 +48,10 @@
     </dependency>
     <dependency>
       <groupId>com.cognifide.aet</groupId>
+      <artifactId>models-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.cognifide.aet</groupId>
       <artifactId>communication</artifactId>
     </dependency>
     <dependency>
