+++ rest-endpoint/pom.xml
@@ -42,6 +42,7 @@
       <groupId>com.cognifide.aet</groupId>
       <artifactId>datastorage-api</artifactId>
     </dependency>
+<!--    todo-->
     <dependency>
       <groupId>com.cognifide.aet</groupId>
       <artifactId>jobs-api</artifactId>
@@ -49,10 +50,6 @@
     <dependency>
       <groupId>com.cognifide.aet</groupId>
       <artifactId>models-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.cognifide.aet</groupId>
-      <artifactId>jobs</artifactId>
     </dependency>
 
     <dependency>
